Add Ukraine mode and make pallet building for ColorCycle modes much more systematic.
Added Ukraine mode that ColorCycles between the two colors of the Ukraine flag.

First attempt at this had a hard to debug problem that was caused by
using too large a value for some of the Ukraine components that
resulted in overflowing 8 bits and irregular colors. This made me come
to terms with the fact that "safe" colors for ColorCycle were
completely non-obvious (even to me) and I added some structure that
scales provided color value to fit ColorCycle's constraints.

Once this structure worked for Ukraine it made sense to adapt all the
ColorCycle modes to it, and tidy color pallet construction generally.

There is a question in mind if color_cycle_pallet_limiter should pick
a scale factor that is used over the whole pallet (v. delegating to
color_cycle_color_limiter to pick a scale factor for each color in the
pallet), but the current approach seems to yield good resutls...and
since this more art than science....

Not unrelated, added some logging.